+++ src/ConnectionConfig.tsx
@@ -1,18 +1,17 @@
-import { DataSourcePluginOptionsEditorProps, onUpdateDatasourceJsonDataOption } from '@grafana/data';
-import { Alert, Field, FieldSet, RadioButtonGroup } from '@grafana/ui';
-
+import { DataSourcePluginOptionsEditorProps } from '@grafana/data';
+import { Alert } from '@grafana/ui';
 import React from 'react';
-import { TEST_IDS } from 'testIds';
-import { JWTConfigEditor } from './components/JWTConfigEditor';
-import { JWTForm } from './components/JWTForm';
+import { AuthConfig } from './components/AuthConfig';
 import { GOOGLE_AUTH_TYPE_OPTIONS } from './constants';
+import { TEST_IDS } from './testIds';
 import { DataSourceOptions, DataSourceSecureJsonData, GoogleAuthType } from './types';
 
 export type ConfigEditorProps = DataSourcePluginOptionsEditorProps<DataSourceOptions, DataSourceSecureJsonData>;
 
 export const ConnectionConfig: React.FC<ConfigEditorProps> = (props: ConfigEditorProps) => {
-  const { options, onOptionsChange } = props;
-  const { jsonData, secureJsonFields, secureJsonData } = options;
+  const {
+    options: { jsonData },
+  } = props;
 
   if (!jsonData.authenticationType) {
     jsonData.authenticationType = GoogleAuthType.JWT;
@@ -20,77 +19,9 @@ export const ConnectionConfig: React.FC<ConfigEditorProps> = (props: ConfigEdito
 
   const isJWT = jsonData.authenticationType === GoogleAuthType.JWT || jsonData.authenticationType === undefined;
 
-  const onAuthTypeChange = (authenticationType: GoogleAuthType) => {
-    onOptionsChange({
-      ...options,
-      jsonData: { ...options.jsonData, authenticationType },
-    });
-  };
-
-  const hasJWTConfigured = Boolean(
-    secureJsonFields &&
-      secureJsonFields.privateKey &&
-      jsonData.clientEmail &&
-      jsonData.defaultProject &&
-      jsonData.tokenUri
-  );
-
-  const onResetApiKey = (jsonData?: Partial<DataSourceOptions>) => {
-    const nextSecureJsonData = { ...secureJsonData };
-    const nextJsonData = !jsonData ? { ...options.jsonData } : { ...options.jsonData, ...jsonData };
-
-    delete nextJsonData.clientEmail;
-    delete nextJsonData.defaultProject;
-    delete nextJsonData.tokenUri;
-    delete nextSecureJsonData.privateKey;
-
-    onOptionsChange({
-      ...options,
-      secureJsonData: nextSecureJsonData,
-      jsonData: nextJsonData,
-    });
-  };
-
-  const onJWTFormChange = (key: keyof DataSourceOptions) => onUpdateDatasourceJsonDataOption(props, key);
-
   return (
     <>
-      <FieldSet label="Authentication">
-        <Field label="Authentication type">
-          <RadioButtonGroup
-            options={GOOGLE_AUTH_TYPE_OPTIONS}
-            value={jsonData.authenticationType || GoogleAuthType.JWT}
-            onChange={onAuthTypeChange}
-          />
-        </Field>
-      </FieldSet>
-
-      {isJWT && (
-        <FieldSet label="JWT Key Details">
-          {hasJWTConfigured ? (
-            <JWTForm options={options.jsonData} onReset={onResetApiKey} onChange={onJWTFormChange} />
-          ) : (
-            <JWTConfigEditor
-              onChange={(jwt) => {
-                onOptionsChange({
-                  ...options,
-                  secureJsonFields: { ...secureJsonFields, privateKey: true },
-                  secureJsonData: {
-                    ...secureJsonData,
-                    privateKey: jwt.privateKey,
-                  },
-                  jsonData: {
-                    ...jsonData,
-                    clientEmail: jwt.clientEmail,
-                    defaultProject: jwt.projectId,
-                    tokenUri: jwt.tokenUri,
-                  },
-                });
-              }}
-            />
-          )}{' '}
-        </FieldSet>
-      )}
+      <AuthConfig authOptions={GOOGLE_AUTH_TYPE_OPTIONS} {...props} />
       <div className="grafana-info-box" style={{ marginTop: '16px' }} data-testid={TEST_IDS.helpBox}>
         <p>
           Don’t know how to get a service account key file or create a service account? Read more{' '}